backend/ocr/requirements_2.txt
google-generativeai==0.8.3
Pillow==10.4.0
groq